+++ README.md
@@ -1,2 +1,5 @@
 campy
 =====
+set FLASK_APP=flaskr
+set FLASK_ENV=development
+flask run
