--- app/templates/index.html
@@ -1,8 +1,0 @@
-<html>
-  <head>
-    <title>campy</title>
-  </head>
-  <body>
-    <img src="{{ url_for('api.get_stream', camera=camera) }}">
-  </body>
-</html>
