+++ flaskr/base_camera.py
