+++ flaskr/cameras/fake/muybridge00.jpg
