+++ flaskr/cameras/fake/muybridge01.jpg
