+++ flaskr/cameras/fake/muybridge02.jpg
