+++ flaskr/cameras/fake/muybridge03.jpg
