+++ flaskr/cameras/fake/muybridge04.jpg
