+++ flaskr/cameras/fake/muybridge05.jpg
