+++ flaskr/cameras/fake/muybridge06.jpg
