+++ flaskr/cameras/fake/muybridge07.jpg
