+++ flaskr/cameras/fake/muybridge09.jpg
