+++ flaskr/cameras/fake/muybridge10.jpg
