+++ flaskr/cameras/fake/muybridge11.jpg
