+++ flaskr/cameras/fake/muybridge12.jpg
