+++ flaskr/cameras/fake/muybridge13.jpg
