+++ flaskr/cameras/fake/muybridge14.jpg
