+++ flaskr/cameras/fake_camera.py
