+++ run.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import sys
-from app import run_app
+from flaskr import run_app
 
 if __name__ == '__main__':
     run_app()
